+++ Braintree-Demo/Services/BraintreeDemoTransactionService.h
@@ -1,8 +1,16 @@
 #import <Foundation/Foundation.h>
 
+typedef NS_ENUM(NSInteger, BraintreeDemoTransactionServiceEnvironment) {
+    BraintreeDemoTransactionServiceEnvironmentSandboxBraintreeSampleMerchant,
+    BraintreeDemoTransactionServiceEnvironmentProductionExecutiveSampleMerchant,
+};
+
 @interface BraintreeDemoTransactionService : NSObject
 
 + (instancetype)sharedService;
++ (BraintreeDemoTransactionServiceEnvironment)mostRecentlyUsedEnvironment;
+
+- (void)setEnvironment:(BraintreeDemoTransactionServiceEnvironment)environment;
 
 - (void)fetchMerchantConfigWithCompletion:(void (^)(NSString *merchantId, NSError *error))completionBlock;
 - (void)createCustomerAndFetchClientTokenWithCompletion:(void (^)(NSString *clientToken, NSError *error))completionBlock;