+++ Braintree/Data/Braintree-Data/BTData.m
@@ -4,13 +4,18 @@
 static NSString *BTDataSharedMerchantId = @"600000";
 
 @interface BTData () <DeviceCollectorSDKDelegate>
+@property (nonatomic, copy) NSString *fraudMerchantId;
 @property (nonatomic, strong) DeviceCollectorSDK *kount;
 @end
 
 @implementation BTData
 
 + (instancetype)defaultDataForEnvironment:(BTDataEnvironment)environment delegate:(id<BTDataDelegate>)delegate {
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     BTData *data = [[BTData alloc] initWithDebugOn:NO];
+#pragma clang diagnostic pop
     [data setDelegate:delegate];
 
     NSString *defaultCollectorUrl;
@@ -29,7 +34,7 @@ static NSString *BTDataSharedMerchantId = @"600000";
             break;
     }
     [data setCollectorUrl:defaultCollectorUrl];
-    [data setKountMerchantId:BTDataSharedMerchantId];
+    [data setFraudMerchantId:BTDataSharedMerchantId];
 
     return data;
 }
@@ -41,9 +46,9 @@ static NSString *BTDataSharedMerchantId = @"600000";
 
         NSArray *skipList;
         if ([CLLocationManager authorizationStatus] == kCLAuthorizationStatusAuthorized && [CLLocationManager locationServicesEnabled]) {
-            skipList = @[DC_COLLECTOR_DEVICE_ID, DC_COLLECTOR_GEO_LOCATION];
-        } else {
             skipList = @[DC_COLLECTOR_DEVICE_ID];
+        } else {
+            skipList = @[DC_COLLECTOR_DEVICE_ID, DC_COLLECTOR_GEO_LOCATION];
         }
         [self.kount setSkipList:skipList];
     }
@@ -54,8 +59,34 @@ static NSString *BTDataSharedMerchantId = @"600000";
     [self.kount setCollectorUrl:url];
 }
 
-- (void)setKountMerchantId:(NSString *)merc{
-    [self.kount setMerchantId:merc];
+- (void)setFraudMerchantId:(NSString *)fraudMerchantId {
+    _fraudMerchantId = fraudMerchantId;
+    [self.kount setMerchantId:fraudMerchantId];
+}
+
+- (void)setKountMerchantId:(NSString *)kountMerchantId{
+    [self setFraudMerchantId:kountMerchantId];
+}
+
+- (NSString *)collectDeviceData {
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+    NSString *deviceSessionId = [self collect];
+#pragma clang diagnostic pop
+
+    NSError *error;
+    NSData *data = [NSJSONSerialization dataWithJSONObject:@{ @"device_session_id": deviceSessionId, @"fraud_merchant_id": self.fraudMerchantId }
+                                                   options:0
+                                                     error:&error];
+    if (error || !data) {
+        return nil;
+    }
+
+    NSString *dataString = [[NSString alloc] initWithData:data
+                                                 encoding:NSUTF8StringEncoding];
+
+    return dataString;
 }
 
 - (NSString *)collect {